Cleanup

+++ app/routes/lists+/.fetch+/delete-watchlist.$request.jsx
@@ -4,6 +4,10 @@ export async function loader(params) {
   try {
     const searchParams = new URLSearchParams(params.params.request);
 
+    if (!searchParams.get('authorization') || searchParams.get('authorization') != process.env.VEUD_API_KEY) {
+      throw new Error("Error: invalid authorization!")
+    }
+
     await prisma.watchlist.delete({
       where: {
         id: searchParams.get('id'),
